functions/line_webhook/requirements.txt
# LINE Bot SDK
line-bot-sdk==3.13.0

# AWS SDK for Python (Boto3) - Latest version for bedrock-agentcore support
boto3>=1.35.70

# HTTP Requests library
requests>=2.31.0
